requirements.txt
dnspython~=2.3.0
motor~=3.1.1
pycryptodome~=3.17
pycryptodomex~=3.17
pymongo~=4.3.3
websockets~=10.4
wincertstore~=0.2

# необходимо для yaml_config, отключите, если используете другую систему конфигурации
pyyaml~=6.0